Web Scraping Project
This program scraps a 'quotes to scrape' website and uploads the data to a CSV file. The game file initiates a new game that reads the data from the CSV file and asks the user to guess who the quote is from. After each incorrect guess, there is a new hint given back.